requirements.txt
asgiref==3.9.1
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==23.0.0
h11==0.16.0
idna==3.10
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
pytube==15.0.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tailwindcss==0.0.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.9.0
